move file match limit test to result package
This moves TestFileMatch_Limit to the package that contains the type it
is actually testing.
