+++ internal/search/result/file_test.go
@@ -289,3 +289,91 @@ func TestChunkMatches_MatchedContent(t *testing.T) {
 		})
 	}
 }
+
+func TestFileMatch_Limit(t *testing.T) {
+	tests := []struct {
+		numHunkRanges       int
+		numSymbolMatches    int
+		limit               int
+		expNumHunkRanges    int
+		expNumSymbolMatches int
+		expRemainingLimit   int
+		wantLimitHit        bool
+	}{
+		{
+			numHunkRanges:     3,
+			numSymbolMatches:  0,
+			limit:             1,
+			expNumHunkRanges:  1,
+			expRemainingLimit: 0,
+			wantLimitHit:      true,
+		},
+		{
+			numHunkRanges:       0,
+			numSymbolMatches:    3,
+			limit:               1,
+			expNumSymbolMatches: 1,
+			expRemainingLimit:   0,
+			wantLimitHit:        true,
+		},
+		{
+			numHunkRanges:     3,
+			numSymbolMatches:  0,
+			limit:             5,
+			expNumHunkRanges:  3,
+			expRemainingLimit: 2,
+			wantLimitHit:      false,
+		},
+		{
+			numHunkRanges:       0,
+			numSymbolMatches:    3,
+			limit:               5,
+			expNumSymbolMatches: 3,
+			expRemainingLimit:   2,
+			wantLimitHit:        false,
+		},
+		{
+			numHunkRanges:     3,
+			numSymbolMatches:  0,
+			limit:             3,
+			expNumHunkRanges:  3,
+			expRemainingLimit: 0,
+			wantLimitHit:      false,
+		},
+		{
+			numHunkRanges:       0,
+			numSymbolMatches:    3,
+			limit:               3,
+			expNumSymbolMatches: 3,
+			expRemainingLimit:   0,
+			wantLimitHit:        false,
+		},
+		{
+			// An empty FileMatch should still count against the limit
+			numHunkRanges:       0,
+			numSymbolMatches:    0,
+			limit:               1,
+			expNumSymbolMatches: 0,
+			expNumHunkRanges:    0,
+			wantLimitHit:        false,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run("", func(t *testing.T) {
+			fileMatch := &FileMatch{
+				File:         File{},
+				ChunkMatches: ChunkMatches{{Ranges: make(Ranges, tt.numHunkRanges)}},
+				Symbols:      make([]*SymbolMatch, tt.numSymbolMatches),
+				LimitHit:     false,
+			}
+
+			got := fileMatch.Limit(tt.limit)
+
+			require.Equal(t, tt.expNumHunkRanges, fileMatch.ChunkMatches.MatchCount())
+			require.Equal(t, tt.expNumSymbolMatches, len(fileMatch.Symbols))
+			require.Equal(t, tt.expRemainingLimit, got)
+			require.Equal(t, tt.wantLimitHit, fileMatch.LimitHit)
+		})
+	}
+}
